Adjusted readme to highlight the buildr.py script
The main script and manifest is why this exists, not the pseudo CI system that is in the agent/coordinator scripts. Devs should be able to fully run the build locally without having to depend on a CI agent being available or getting "red builds" while they test out new ways to build their code.